% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_season.R
\name{get_current_week}
\alias{get_current_week}
\title{Get Current Week}
\usage{
get_current_week()
}
\value{
current nfl regular season week as a numeric
}
\description{
A helper function that returns the upcoming NFL regular season week based on
date the number of weeks since the first Monday of September (Labor Day).
This will count a new week starting on Thursdays.
}
\examples{
get_current_week()

get_current_week()
}
