% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_season.R
\name{get_labor_day}
\alias{get_labor_day}
\title{Compute Date of Labor Day}
\usage{
get_labor_day(year)
}
\arguments{
\item{year}{Numeric or Character year. 4 Digits.}
}
\value{
Object of Class \code{Date}
}
\description{
Computes the date of the Labor Day, i.e. the first Monday in September, in a given year.
}
\examples{
# 2023 Labor Day was on Sep 4th
compute_labor_day(2023)
}
