% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_season.R
\name{get_season_range}
\alias{get_season_range}
\title{Compute Range of Seasons}
\usage{
get_season_range(start = 1999, end = get_current_season())
}
\arguments{
\item{start}{Numeric or Character year. 4 Digits.}

\item{end}{Numeric or Character year. 4 Digits. Defaults to \code{\link[=get_current_season]{get_current_season()}}}
}
\value{
Object of Class \code{integer}
}
\description{
Computes a range of seasons from given start to end. This is intended to be
used in workflow setups when code should run for multiple seasons
}
\examples{
# Compute range from 2016 to current season
get_season_range(2016)
}
