% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{get_teams}
\alias{get_teams}
\title{Get NFL team abbreviations}
\usage{
get_teams(as_json = !interactive())
}
\arguments{
\item{as_json}{If \code{TRUE}, the function will output in JSON format. Defaults
to TRUE in non-interactive sessions.}
}
\value{
An array of team abbreviations (character)
}
\description{
Get NFL team abbreviations
}
\examples{
get_teams()
get_teams(as_json = TRUE)
}
